+++ client/src/features/MovieDetails/index.js
@@ -1,2 +1,2 @@
-export { default } from './container.connected';
+export { default } from './container.enhanced';
 export { default as reducer } from './reducer';
