+++ client/src/features/SearchForm/index.js
@@ -1,1 +1,1 @@
-export { default } from './container.connected';
+export { default } from './container.enhanced';
